V1.3.1- Updated integer types
-Added stdint definitions and clarified RAM considerations

+++ Well_Tester/Communication_Functions.h
@@ -1,10 +1,12 @@
+#include <stdint.h>
+
 #ifndef COMMUNICATION_FUNCTIONS_H
 #define COMMUNICATION_FUNCTIONS_H
 
 void Setup_Comms(void);
 void Print_String(char *);
-void Print_UINT(unsigned int Value);
-void Print_ULONG(unsigned long Value);
+void Print_UINT(uint16_t Value);
+void Print_ULONG(uint32_t Value);
 void Output_Result(void);
 
 #endif
