+++ Well_Tester/cstbool.h
@@ -1,8 +1,9 @@
+#include <stdint.h>
 
 #ifndef CSTBOOL_H_
 #define CSTBOOL_H_
 
-typedef int bool;
+typedef uint8_t bool;
 #define true  1
 #define false 0
